Use go's new dep tool to manage dependencies
Go's official dependency management too is going to be dep [1], I have
used this to lock the dependencies for this project down. One cool feature
that I really like is that I can change the source of a dependency.

Although we are using github.com/xanzy/go-gitlab to interact with gitlab,
There are some custom additions that have yet to be merged upstream. We
can tell dep to pull from our fork, yet keep the references the same. This
Allows us to update very easy should things get merged upstream :-)

[1]: https://github.com/golang/dep

Signed-off-by: Gavin-John Noonan <Gavinjohn.noonan@gmail.com>